[fix] decreased performance
When inserting a temporary object into a container using the methods `push_back`, the object is constructed outside the container and then moved/copied into the container. On the other hand, the `emplace_back` methods allow you to eliminate the extra call of the move/copy constructor and create the object "in place" inside the container instead, perfectly passing the function's parameters to the constructor.